+++ rag/package.json
@@ -9,24 +9,19 @@
     "typecheck": "tsc --noEmit",
     "format": "prettier --write .",
     "format:check": "prettier --check .",
-    "agent": "tsx src/agent.ts",
-    "openai-agent": "tsx src/openai-vector-agent.ts",
+    "agent": "tsx scripts/openai-vector-agent.ts",
     "psql": "psql emailsmart_rag"
   },
   "dependencies": {
+    "dotenv": "^17.1.0",
+    "openai": "^5.8.3"
+  },
+  "devDependencies": {
+    "@openai/agents": "^0.0.10",
     "@langchain/community": "^0.3.48",
     "@langchain/core": "^0.3.62",
     "@langchain/openai": "^0.5.18",
     "@neondatabase/serverless": "^1.0.1",
-    "@openai/agents": "^0.0.10",
-    "@openai/agents-openai": "^0.0.10",
-    "dotenv": "^17.1.0",
-    "drizzle-orm": "^0.44.2",
-    "openai": "^5.8.3",
-    "pg": "^8.16.3",
-    "zod": "^3.25.76"
-  },
-  "devDependencies": {
     "@eslint/js": "^9.30.1",
     "@types/node": "^20.19.5",
     "drizzle-kit": "^0.31.4",
@@ -36,7 +31,9 @@
     "prettier": "^3.6.2",
     "tsx": "^4.7.1",
     "typescript": "^5.8.3",
-    "typescript-eslint": "^8.36.0"
+    "typescript-eslint": "^8.36.0",
+    "pg": "^8.16.3",
+    "zod": "^3.25.76"
   },
   "overrides": {
     "@browserbasehq/stagehand": {
